+++ src/verify-release.js
